@@ -14,24 +14,24 @@ classifiers = [
   "License :: OSI Approved :: MIT License",
   "Programming Language :: Python",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
   "Framework :: AsyncIO",
   "Framework :: Trio",
 ]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 
-asyncpg = {version = "*", optional = true}
-aiopg = {version = "*", optional = true}
-aiomysql = {version = "*", optional = true}
-aiosqlite = {version = "*", optional = true}
-aioodbc = {version = "*", optional = true}
-triopg = {version = "*", optional = true}
-trio_mysql = {version = "*", optional = true}
+asyncpg = { version = "*", optional = true }
+aiopg = { version = "*", optional = true }
+aiomysql = { version = "*", optional = true }
+aiosqlite = { version = "*", optional = true }
+aioodbc = { version = "*", optional = true }
+triopg = { version = "*", optional = true }
+trio_mysql = { version = "*", optional = true }
 
 [tool.poetry.extras]
 asyncpg = ["asyncpg"]
@@ -52,7 +52,7 @@ cryptography = "*"
 psycopg2-binary = "*"
 pypika-orm = "^0.0.11"
 pytest = "*"
-pytest-aio = {version="^1", extras=["trio"]}
+pytest-aio = { version = "^1", extras = ["trio"] }
 pytest-mypy = "*"
 trio_asyncio = "*"
 trio_mysql = "*"
@@ -62,7 +62,7 @@ ruff = "*"
 pre-commit = "*"
 
 [tool.pytest.ini_options]
-addopts = "--strict-markers --mypy -xsvl"
+addopts = "--strict-markers --mypy -xsvl tests"
 
 [tool.mypy]
 packages = ["aio_databases", "tests"]
@@ -72,7 +72,7 @@ check_untyped_defs = true
 [tool.tox]
 legacy_tox_ini = """
 [tox]
-envlist = py38,py39,py310,py311
+envlist = py39,py310,py311,py312
 isolated_buiid = true
 
 [testenv]
@@ -87,20 +87,37 @@ commands =
 
 [tool.ruff]
 line-length = 100
-target-version = "py38"
+target-version = "py39"
 exclude = [".venv", "docs", "examples"]
+
+[tool.ruff.lint]
 select = ["ALL"]
 ignore = [
-  "ARG", "D", "UP", "ANN", "DJ", "EM", "RSE", "SLF", "RET", "S101",
-  "PLR2004", "PLR0912", "N804", "A003", "TRY003", "COM812", "PT012"
+  "ARG",
+  "D",
+  "UP",
+  "ANN",
+  "DJ",
+  "EM",
+  "RSE",
+  "SLF",
+  "RET",
+  "S101",
+  "PLR2004",
+  "PLR0912",
+  "N804",
+  "A003",
+  "TRY003",
+  "COM812",
+  "PT012",
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "test*.py" = ["FBT001"]
 
 [tool.black]
 line-length = 100
-target-version = ["py38", "py39", "py310", "py311"]
+target-version = ["py39", "py310", "py311", "py312"]
 preview = true
 
 [build-system]
